Updated Pre-Installation to handle uninstallation of previous version better
Some software like NI and STK include connectors for MATLAB that will show up when Get-InstalledApplication is run. If the format of the uninstall string doesn't match our script, it errors out. We want to make sure the script only grabs previous versions of MATLAB and leaves connectors and interfaces alone.
- The uninstaller in Pre-Installation now checks to see if the uninstall string is empty before parsing it.
- The uninstaller in Pre-Installation checks to see if MATLAB and uninstall.exe are part of the returned uninstall string before proceeding.